deep-learning-week/ml-development/interview-pipeline/requirements.txt
openai
pdfplumber
python-docx
python-dotenv
